// pronti-frontend-main/front/src/app/modulos/empresa/form/form-empresa.enum.ts
export enum FormEmpresaEnum {
  // nombre = 'nombre',
  nombreComercial = 'nombreComercial',
  razonSocial = 'razonSocial',
  rucEmpresa = 'rucEmpresa',
  direccionEmpresa = 'direccionEmpresa',
  telefonoEmpresa = 'telefonoEmpresa',
  documentoRepresentanteLegal = 'documentoRepresentanteLegal',
  nombreRepresentanteLegal = 'nombreRepresentanteLegal',
  nombreContador = 'nombreContador',
  rucContador = 'rucContador',
  telefonoContador = 'telefonoContador',
  correoEmpresa = 'correoEmpresa',
  correoContador = 'correoContador',
  correoRepresentanteLegal = 'correoRepresentanteLegal',
  tipoEmpresa = 'tipoEmpresa',
  claseContribuyente = 'claseContribuyente',
  obligadoLlevarContabilidad = 'obligadoLlevarContabilidad',
  agenteRetencion = 'agenteRetencion',




}